requirements.txt
django
mysqlclient
django-filter
djangorestframework
django-rest-swagger
django-debug-toolbar
